Change to use CMake
