+++ .gitignore
@@ -1,0 +1,4 @@
+**/*~
+latex
+html
+build
