+++ resources/icon.qrc
@@ -1,0 +1,5 @@
+<!DOCTYPE RCC><RCC version="1.0">
+<qresource>
+	<file>images/icon.xpm</file>
+</qresource>
+</RCC>
