+++ resources/images/icon.xpm
@@ -1,0 +1,75 @@
+/* XPM */
+static char * icon_xpm[] = {
+"50 50 22 1",
+" 	c #366972",
+".	c #488AAA",
+"+	c #3F3F3F",
+"@	c #E1E1E1",
+"#	c #F2F2F2",
+"$	c #C4C4C4",
+"%	c #C5276F",
+"&	c #FB7D91",
+"*	c #9A8F7E",
+"=	c #D2D1BF",
+"-	c #B9B89B",
+";	c #62480D",
+">	c #DDBF17",
+",	c #E4A93A",
+"'	c #EEF786",
+")	c #ECDF3C",
+"!	c #8B8B8B",
+"~	c #EBEBE7",
+"{	c #FFFFFF",
+"]	c #333333",
+"^	c #313131",
+"/	c #555555",
+"      ........................................    ",
+" ++++++  .......................................  ",
+"++@##@+++........................................ ",
+"$$$##$$$+...................................%%... ",
+"$++++++$+................................. %&&%...",
+"++    +++................................ %&&&&% .",
+"  ....   ................................ %&&&&&% ",
+" ....................................... %&&&&&&% ",
+" ++++++  ...............................+++&&&&&% ",
+"++@##@+++.......                        +$#++&&&% ",
+"$$$##$$$+....**************************+$##@@+&%  ",
+"$++++++$+....*========================-++##@$++% .",
+"++    +++... *========================-;>+@@$$+  .",
+"  ....   ... *=======================-;>,'+$$$+ ..",
+" ........... *=======================-;,'),+$+ ...",
+" ++++++  ... *======================-;>,),>;+  ...",
+"++@##@+++... *======================-;,'),>;;  ...",
+"$$$##$$$+... *=====================-;>,),>>;*  ...",
+"$++++++$+... *=====================-;,'),>;-*  ...",
+"++    +++... *====================-;>,),>>;-*  ...",
+"  ....   ... *======%=============-;,'),>;--* ....",
+" ........... *=====%&%============;>,),>>;--* ....",
+" ++++++..... *======%============-;,'),>;--=* ....",
+"++@##@+++... *======!===========-;>,),>>;-==* ....",
+"$$$##$$$+... *=!!!!!!!!!!!======-;,'),>;--==* ....",
+"$++++++$+... *=!#########!======;>,),>>;--==* ....",
+"++    +++... *=!#########!=====-;,'),>;--===* ....",
+"  ....   ... *=!#########!====-;>,),>>;-====* ....",
+" ........... *%!#!!###!!#!%===-;,'),>;--====* ....",
+" ++++++  ... *%!#!!###!!#!%===;>,),>>;--====* ....",
+"++@##@+++... *%!#########!%==-;,'),>;--=====* ....",
+"$$$##$$$+... *=!###!#!###!==-;>,),>>;-======* ....",
+"$++++++$+... *=!####!####!==-;,'),>;--======* ....",
+"++    +++... *=!!!!!!!!!!!==;>,),>>;-=======* ....",
+"  ....   ... *=====!#!=====-;,'),>;--=======* ....",
+" ........... *=!!!!!!!!!!!=;;;;;>>;--=======* ....",
+" ++++++  ....*=!#########!-;~{{;>;--========* ....",
+"++@##@+++....**!!!!!!!!!!!;~{{{;;;*********** ....",
+"$$$##$$$+....             ]~{{~~;;            ....",
+"$++++++$+..............   ]^+~~;;     ............",
+"++    +++................ ^^^^];   ...............",
+"  ....   ................ ]/^]]  .................",
+" .........................]^]] ...................",
+" ++++++  .................]]] ....................",
+"++@##@+++.........................................",
+"$$$##$$$+.........................................",
+"$++++++$+.........................................",
+"++    +++........................................ ",
+"    ..   ......................................   ",
+"     ........................................     "};
